Proof of Concept
This is a mostly-working prototype for the idea.

You can press F2 to activate the kill a beetle achievement
Then when you kill a beetle, you get a notification telling you that you
completed the achievement and you get an item as a reward.
Then you can press F2 to activate the achievement again.

Only TasksPlugin and TempAchievements actually do anything
The rest of the files are from my first attempt to structure the tasks.
The plan is to refactor them a little bit to turn them into the version of
achievements I ended up using.
Hopefully it's as easy as changing some names of methods and extending
from the right class.