Rewards
Air Kills
Gave the task a readable name for popups and in the logbook
Task can be cancelled (like if someone finished it before you)
id is now an enum. Seemed easier than trying to keep track of a number

TasksPlugin
Made rewards
Create rewards that map to a certain task when the tasks are created.
Can give a specific item or temp items.
Temp items are items you only have for the current stage
Gold and xp rewards aren't hooked up yet
Generates tasks on start of the run instead of on F2 (well, in addition
to)
Removing the temp items on command works mostly.
The problem is the wiki says CharacterMaster.readOnlyList is a list of
players, but it is actually just a list of everything (players and mobs)
So I just need a different way to store the players. I think I can test of
a CharacterMaster has a playerCharacterMasterController
Need to rework the way items are randomly selected and given. It's too
random. You can get higher tier items too easily and items that don't
really exist. And I think I use a variety of different methods that all do
the same thing/duplicate code when I should make a method.
ItemDropAPI might help.