Multiplayer Testing
Did not survive the first contact with another player.
At least I learned some stuff.

Actions that I need to use to track the achievements don't run on the
client. Like onCharacterDeathGlobal

Tried some ClientRpc and Command unity networking stuff that didn't work
either.

Made a list to hold only players, not mobs which was a plus. Not that it's
100% implemented where it should be yet. It is used to remove temp items.
I'm not sure if it should just be a list of CharacterMasters or if it should be a struct with some more data like player number, etc.