Server Tracking
Server now does everything. Pretty much
Each client sets up their achievements, but they don't really do anything.
It's just to make it easy on me to get them to do popups.
The server hooks up the achievements to actually track stuff, then the
server tracks each player's stats in the achievement and registers it as
completed when the first player completes it

AirKills
Now tracks an array of kills instead of just one set
Got the event for when someone hits the ground working I think (haven't
tested it in multiplayer).
Should work for any number of players and it can unsubscribe itself if I
save the delegates to a list. Now that I think about it, I don't remove
them from the list. It should still work; the list will keep getting
longer, but it will only look at the first numPlayers elements.

Changed a lot of methods in Tasks and AirKills to work for a specific
player.

Other Task children:
Just edited them so they don't throw errors.

TasksPlugin
Most of it only runs on the server.
Update could be a source of some bugs by having it run on the client, but
it's only there for debugging anyway.
Swapped out netID for player number.
Has helper methods for grabbing specific player's CharacterMaster
components. They are used by the achievements to track stats, etc.

RpcTaskCompletion
Is who knows. Maybe it works? It's to create the popup on the client's
screen. localPlayer sounds like it would work, but I feel like in the
testing I did a little while ago, it was always false (or it was a
different similar variable I'm thinking of)

I need something to run on Awake where a client can store something unique
about them to be able to figure out what player number they are. The
server just sends out playerNum finished this task. The client should be
able to figure out if that playerNum refers to them. Hopefully
isLocalPlayer works, but who knows?