Clientside Popups
The server can now tell the client that it completed a task and should
therefore show the achievement popup.

Unity's Rpc system didn't work for me so I used miniRpc from the modding
wiki

AirKills
You get an error if you try to do hook/unhook things on a dead player. I
think it's because they don't have a body or motor anymore
So I'm just skipping players if they don't have all the right components.
This skips the unsubscribe step.  I'm hoping that it doesn't matter. Like
when the stuff goes null, it also empties onHitGround so I don't have to.

TaskPlugin
Made IRpcActions that allow me to run stuff on the clients and send data.
taskCompletionClient seems to work. You get popups.
updateTasksClient is untested. Can I send a whole array of ints? We'll
find out some day.
Do these actions need to be set up on the client and server? Should they
be after a if(serverActive)?
When tasks are randomly created, add those tasks to a list to send to the
clients. They should also be updated as people complete tasks or the level
ends or more tasks become available.
Other clients aren't notified when someone completes a task. Only the
client that completed it. Maybe updateTasksClient will solve that problem.

To Do:
Make more tasks
Tell people what tasks are active and completed
Do that, but put it in a pretty UI
Remove code that isn't needed anymore

Bugs/Issues:
I'm not sure if player numbers are synced. I use NetworkUser and
CharacterMaster.readOnlyList. Do they both have players in the same order?
Hard to test when you're using a host and client. The host is probably
always player 0, but what happens when you have multiple clients?