Starting UI
Figured out how to add my tasks to the objectives list on the side (ther
red box where it says "find tp", etc
In trying to do that, I figured out how to make generic notifications
(like when you pickup an item)
They may be a better candidate for telling you when you complete a task
(instead of the achievement UI)

Tasks
Made description static so I can access it easily

TaskPlugin
UpdateTasksClient RPC action was giving me trouble. It wouldn't let me
pass an array like I wanted. I had to switch to sending tasks one at a
time. It sends some extra data like how many tasks the server is sending
and if they're completed (to be used to update players when one player
completes a task)

I removed a bunch of commented out code and that messed up the diff a bit.

I can add some generic text to the objectives UI fairly easily, but I
don't know what the limitations are. It just gives me a checkbox, but I
don't know if I can interact with it.
All I'm doing now is hiding my task UI objects when they're complete. I'm
not sure if I can tie into RoR's system where they get a strikethrough and
fade over time. So some remaining code is my first attempts at that that
haven't panned out yet.

To add to the objectives panel, all I had to do was instantiate the same
prefab as they did and parent it like they did and it just auto adjusts
for me. I thought I would have to do the spacing, but it just worked. It
even seems to work for when the official objectives change. Like when you
activate the tp, "find tp" goes away and "kill boss" shows up and the
official stuff stays above my stuff (which I want)

I'm still not sure how to hook up to anything other than
Run.onRunStartGlobal.
I need to find onStageStart and onTeleActivated. Might have an idea how to
do the tele stuff. Leaving a note to find it later

I'm not sure what the best way to show completion of tasks in the list.
All I have now is hiding the complete ones.

Notifications
Their position is in pixel coords it seems (0, 0) to (1920, 1080)
Some of these "later" methods were from when I was trying to figure that
out. I thought it was a (0,0) to (1,1) scale so stuff wasn't moving and I
thought it was a timing thing. Like I had to either move before awake or
after awake or before/after parenting
Can probably be deleted

GetTaskDescription
current strategy is to just have a switch() for each different task to get
their description. Couldn't think of something better.

TaskInfo
Stores some info about that tasks that I can send over the network. index
and total are to help with sending a whole list of tasks so the client
knows which order to draw them in and how many to expect. Makes it so I
don't have to search the arrray for a match or resize the array as more
tasks show up

To Do
Now I just need more tasks to be able to test multiple tasks
And ??? hidden tasks that unlock at a certain time. Also work as default
tasks for when I haven't made 9+ tasks