Second Task, Better Reset
AirKills
Figured out which fields in the XML need to be deleted for the
achievements to be trackable again
Removed some bloat commented out code

DamageMultiple
Added this task
The class has been there since version 0.1, but updated it to work with
the current way of tracking the tasks.

Task
Tested some other versions of onRunDesetroyGlobal, but they didn't seem to
ever trigger. I fixed the issue of tasks not resetting after you got them
elsewhere.

TasksPlugin
Renamed a global variable that had the same name as some local variables
in some methods
HasAchievement += this is how I reset my achievements
Remove the === Client ==== Server === comments that I don't think were
even right
I properly store a copy of the UI that I make and check if they exist
first before making a new one. Important bc when I get a new task, it
iterates over all tasks to update the UI
Made a helper method to create random unique tasks. Current system can't
accomodate multiple tasks of the same type (like get 3 airborne kills and
get 5 airborne kills)
Moved notification code to its own method NotificationTest. It never gets
called. It's still alive so I can check it to find out how to make
notifications again.

TaskCompletion() was getting called twice. Thought it was getting called
from the client and from the server. Nope. turns out in awake, I don't
need to call both airkills.onCompletion += TaskCompletion and
damageMulti.onCompletion += TaskCompletion.
OnCompletion is static and just passes the current task as a parameter. so
doing both += is just like doing
task.onCompletion += TaskCompletion
task.onCompletion += TaskCompletion
Which is not what I want. So for new tasks, all I have to do is
UnlockablesAPI.AddUnlockable<newTaskClass>(true)
Not even sure if the true is necessary.

To Do
I might need to restructure some stuff. Tasks plugin is getting rather
long. It could probably be split up into multiple classes. It's gotten
bloated for a number of reasons.
Not knowing how the mod structure works (does everything need to be in the
same class to run properly, if I have multiple classes, how does it work
sequencially, where is the entry point, etc.) not really, entry point is
the class that extends BaseUnityPlugin
The code base to learn from is so large that it's hard to find what I'm
looking for so when I find something that might be useful shortly, it's
useful to just copy a piece of it so I have it when I'm ready, but it
takes up space.
The slow turnaround on testing. Running the game takes a while so it's
hard to test so it's hard to tell what can be deleted if I can't test edge
cases very well. Like I'm 90% sure I can remove the notification move
later stuff