StayInAir and DamageInTime
Updated these tasks to work now

DamageInTime
Task classes do not extend monobehaviour so they can't start coroutines.
So it has the coroutine run on TasksPlugin.
Whenever you do damage, add to the total and set a coroutine to run in 5
sec to decrement that same amount.

StayInAir
Was going to use the hit ground event system from AirKills, but that just
tracked when you hit the ground, not when you left the ground.
So I looked up how the merc's stay in the air achievement worked and just
copied that mostly
Bugs?
It's possible this breaks when one of the players dies. I think their body
and motor go away so trying to use them might break this.

TasksPlugin
instance to allow DamageInTime to run coroutines
Removed a bunch of old code that wasn't being used.