Chest Items and Tele Actions
Found some actions to sub to to figure out when stuff happens
I have:
When you interact with the tp, when the TP event starts, when you charge
it to 100%, when you kill the boss, how fast you killed the boss, and when
you click on the tp again to go to the next map
And maybe when the next map loads. The UI seems to be active, not sure
about players yet

Rewards
Got the game's random items systems to work for me.
Like the different chests already have different loot tables. So now I can
simulate opening a specific (or random) chest to get a semi-random item,
but not get unimplemented items. And the rarity works such that white
items are more common than green, etc.