Mutliplayer Bug Fixes and Testing
Game seems to work for multiplayer. You can all compete to complete tasks
and get new tasks on new stages (same 4 tasks for now)
Probably some desync bc the clients are calling RemoveItems and StageEnd
when the players leave the stage. Forgot to fix that.

Task
if you pass in -1 to deactivate, it will force it to deactivate
(otherwise, the parameter passed in has to match)

TasksPlugin
taskEndedClient - this is called on each client when a task is completed
so they know to remove it from the UI

OnDeactivate(-1) deactivate all tasks at the start of the game so you
don't still have active tasks from the previous game.

Removed all of update. It was just old testing stuff that doesn't really
work anymore anyway

Rewards
Swapped to using PickupIndex instead of ItemIndex. I needed a pickupIndex
for something and you can go pickup -> item, but not the other way around.
Also, itemIndex seems to be depreciated so I tried to move away from it,
but some things like player.inventory.giveItem(ItemIndex) still need item
index so it didn't really help.

Chat.AddPickupMessage doesn't push the message to other clients like I
thought it would.
I think I need more miniRPC client methods to send the clients for them to
create their own pickup messages. Or rework the way I give out items.

Removed a bunch of old comments that are unhelpful