New Interaction Tasks
Figured out how to tell when a player opens a chest, barrel, printer, etc.
Then implemented the tasks I had that required those sort of things.
Plus farthestAway. Whomever is the farthest distance away in 20 seconds
wins.

TasksPlugin
the code that's like
go?.GetComponent<>
was just the testing bits. None of it really matters. Committing it so
maybe I can reference it later if I delete it and forget.