Reward Icons
Tasks on the side now show what item you'll get as a reward
Made the reward serialize so it can be sent to clients.
(this is the only important part of this commit)

TasksPlugin
Track when a skill is used.
Made tasks for this too. DontUseSkill triggers when all but 1 player uses
a skill, whomever hadn't used it gets the reward. If multiple go the whole
stage without the skill, they both get the reward. Doesn't work
singleplayer.
Ordered skill is first to use their skills in order wins. It's a little
finnicky. engi and huntress press m1 as part of their R skills, but that
causes their m1 skill to register.
These skill tasks aren't as interesting as I thought they'd be.

Hide the checkbox, put an item icon in its place. Seems to work pretty
good and looks perfect too

Tested giving out items as pickup orbs. Problem is they can't be picked up
until they hit the ground. I wanted them to spawn inside the player and be
picked up immediately so the game handled the "player1 picked up item"
messages.
Might still work for a task where you steal another player's items.