Custom Notifications
When you complete a task, a notification pops up beside the task to tell
you that you completed it.

itemIconPrefab
last update, display.itemIconPrefab worked. Now it's null at GameSetup. So
it also tries to find the prefab when it creates the task objective UI
objects. Is it null because of a small timing change? I used to create
dictionary before that, but deleted that code bc I never used the dict.
Did that give the field time to fill in? Kinda doubt it

When you complete a task, instead of jsut hiding the UI element, it gets a
strikethrough and everything goes grey instead. I might have to do that
when someone else completes a task... Not sure if it works for that.

CreateNotification()
Used the base notification, but scaled it smaller and made the text
bigger. Used trial and error to get it to appear near the appropriate task
UI. So it's not really perfect, but it works. Need to do a better job on
what the text actually says. It's not super readable to tell what task you
completed.