Not Achievements Anymore
Task used to extend Achievement. Now it doesn't
Removed some strings from each task and changed their type to be public
The only thing that really changes is the achievement popup at the top of
the screen doesn't happen when you complete a task anymore, but as of last
commit, I made notifications pop up instead.

TaskPlugin
Need to setup tasks differently.
The task objects need to be created when a real game starts (spawn a
character on the map). The old version created them when the game launched
to the main menu.
I also store the task objects so I can access them. I don't know if I
could access them with the old system. If there was a way, I didn't know
it

DealDamageInTime
Started testing what it would look like to have tasks scale as the game
goes on and to have descriptions that reflect the current goals.