Multiplayer Flagged Tasks
Check if a task can be activated before you activate it
Turns some tasks off if not in multiplayer.

GetDescription
Can get the right description for tasks
Tasks can have variable descriptions if their goals change based on the
difficulty
I may have done this in the weirdest way. The old system was kinda weird
too. This way works and seemed like it would be the least copy/pasting
The new keyword messed up the base case

CanActivate
Could I have made this in such a way that it wasn't just copy and pasting
for each task that was multiplayer only?
Might have to revisit this if I have other flags that affect multiple
tasks.

StayInAir
Scales with difficulty coefficient