Progress UI
Made a bar to go below the task description to show how far along you are.

Each Task needs to update this progress info at different points that are
speciifc to that task
I've only implemented the easiest ones

TasksPlugin
Create 2 new UI objects to show the progress of you and a rival. Either
whomever is in first or if that's you, whomever is in second.
Finally figured out how to add UI elements. They just sit on top of the
existing objective panel stuff. I had hoped I could add them under each
task so I could have more room, but I couldn't figure out how to do that.
UpdateProgress() is what sizes the progress bar. The varaiant with the
extra bool is for use with the rival progress bar to control its colour
and depth.
UpdateTaskProgress. Triggered by the tasks to tell the server when it
should tell the clients to update progress.
These names are simiar and maybe should be changed

Tried a new way of handling getting items from tasks. The hope was that
this would handle sending the info to all players when someone picks
something up. It's also an all-in-one chat message and notification which
previously took up 2 lines.

Removed HasAchievement as I don't use achievements anymore
GetTaskDescription could also be simplified now that I'm storing my own
copies of tasks instead of letting the achievement stuff handle it

ProgressInfo
Info for sending data to clients about progress.
I don't know how to send floats so I convert to an int and then convert
back. I only retain 1 decimal place, but that's all the fidelity I need on
progress anyway.

Haven't tested anything in multiplayer to see if the rival stuff is
working.

CreateTaskProgressBar and UpdateProgress
Seem like they could be moved to a UIHelper class of some sort. I want to
try to figure out what kinds of different progress UI I might need for
diffrent tasks to see what the class structure could be.
Like are there a number of different types of UI that could warrant a
class and sub classes or just 1-2 random methods to help  1-2 random
special tasks?