Baby Drone
Made a task to keep a drone alive. It spawns in with 1hp

The Task spawns a drone interactable, then auto buys it. This spawns the
real drone so I store that drone. Then I change its name so you can tell
which drone needs to stay alive, and finally it deals damage to the drone
1 seconds later. It seems the health isn't ready to take damage right away
(or it just resets the health after it takes the damage)

The balance of this task might be off (but so are all of them really).
If you get the drone at lvl 1, it doesn't regen health. At higher levels,
it does (but without the base regen). It should still be significant.
LevelRegen is 0.2 * baseRegen. Lvl 5, you're half as effective, Lvl 10,
2/3 as effective

I can use the same system to spawn any objects I think. It's the system
the game uses to spawn the lockbox.
I could also use this system to give drones as rewards.
Or chests for reduced costs (or 0). Could be handy for fireworks

Fixed an issue quitting and starting new games.
Things weren't getting unhooked.
Seems to be working now