Command Reward
You can get a command orb as a reward.
Made Die and FindLockbox tasks

BabyDrone
Made a bunch of different names for your drone and different death
messages.
Found a new way to potentially send messages to each player. So maybe
they'll all see the death message

OrderedSkills
Updated the description

Command Reward
To get a command orb: turn on the command artifact, make an item droplet.
When it hits the ground, it turns into a command item. Then I can turn
command off.
It makes the next item to hit the ground into the command droplet. Rarely
it could mess up if another item hits the ground first.
The UI for this is also a little weird. It shows an item with the command
artifact in the bottom left. The item is just there to tell you what tier
the command orb will be. If the bug from ^ happens, that's the item that
will drop. Changing the colour of the command icon doesn't really work.
Maybe the item scrap icons could be a better alternative than a random
item.

More Bugs
If you're playing with the command aritfact on, it will turn it off as
soon as you get a command reward.