HealingItem and NoJump
New task to find a healing item
New task to not jump (only works for multiplayer)
Basically the same as don't use ability

When the lockbox task starts, everyone is given a key so they can open the
lockbox if they find it.
They aren't taken away after. Probably not a big deal.