Config
Added a config to control the weights of tasks

ConfigManager
Makes the config and grabs the values from it. TaskPlugin calls this to
check what the weights are.

TasksPlugin
Passes the config file to ConfigManager. This class is the baseUnityPlugin
so it has the reference to the config file
GetWeightedTasks
Might work. Not hooked up or tested. Uses the weights intead of just
random tasks.
I do feel like evluating the weightSelection could be simplified. Am I
doing redundant stuff?