Multiplayer Testing
First test, multiplayer didn't work. Got null errors on the client
(everything worked fine on server)
I figured the issue must be with miniRPC so I swapped to R2API
NetworkingAPI
But then I still had the same issue.....
Issues ended up being trying to use arrays that were null.
The code that made the arrays ran on the server code, but not on the
client.
I must have goofed it when I made tasks only generate on stages with the
tele bc that's where the problem was.

ServerOnTelePlaced()
telePlaced = true

ClientCode()
if(tele)
array = new [100]

clientCode()
array[i] = thing

Just have to do:

ClientCode()
array = [100]
if(tele)
stuffThatActuallyCaresAboutTele

The old version of the miniRPC I have already figured out how to send a
message (like you are the winner of this task) to one specific player.
NetworkingAPI sends all messages to everyone.
I tried to also send the player number and have the clients check if it
was them, but it didn't work. So I either need a different token to send
so the client knows it won or need to send the ordered list of players
from the server instead of each client building their own. That might
work.
But I'm pretty sure miniRPC works now.

I think these work for the server and clients now
Server can send arbitrary chat messages to everyone
You get a popup when you win
Completed tasks are stroked out

Multiplayer is hard to test solo.