QuickDraw
New Task: QuickDraw
Kill something within 3s of it spawning.
Not 100% if it works on stage 2
3s seems good for lemurs and prob wisps. Beetles might be alive when
underground. Might be too tight at 3s. Probably not a big issue.

Bugs:
Quitting to menu and updating progress.
Seems the rect is null and is trying to be updated. But rect is null is
false.
Causes some things to not get unhooked

KillStreak
Haven't figure out how to ignore REX's self damage.
Did figure out how to ignore fall damage if I want to.