FarKill
New task
Kill something far away.
If you're melee, you can leave the ground and fly towards them and it
counts the distance from your last ground point and the enemy.
This also works for ranged survivors. It's so it's possible for melee, but
it works for both.

Should work. Distance might be way off.
Whoops. Forgot. Still need to update progress with distanceOfLastKill/distanceYouNeed