Unhook and StageEnd split
Split unhook() into unhook() and StageEnd() for a bunch of tasks
Some tasks end at the end of stage. Tasks like 'get the most kills'. And
at the end, all active tasks get unhooked. So I made the sloppy choice to
just combine those two things. That led to some bugs. Sure, tasks can end
at the end of the stage, but there are multiple reasons why you would want
to unhook a task. And reasons you would want to unhook a task without
finding a winner. i.e. cancelling it. Like when everyone dies. So I
separated the functionality of unhooking dependencies and evaluating a
winner like it should have been from the start

On.RoR2.AllPlayersAreDeadUnhookStuffToAvoidBugs
There's a bug when you all die and go to the stats screen. Maybe it's
fixed? Trying to find an event that fires at the rigth time is a bit
tricky. The ones I found might work, but I haven't tested in multiplayer
yet.

Custom win messages
Say in chat who won what task and how they won it (their score, basically)
Each task can set a custom win message specific to that task like A won
with 5 kills or B won by finding 6 chests.
Can set custom colouring for the different parts of the messages.
Double messages for now. I'm not sure which messaging method makes it show
up on all clients. So that needs to be tested in multiplayer.
I'm not sure if these messages even solve the issue of not knowing who
won. The chat box is so small and it dissappears too quickly that you'll
probably miss the message all the same.

Removed some old commented out code from a bunch of tasks.
TasksPlugin needs a trimming like that so much more, but it's too much of
a mess that I can't tell what is and is not a load-bearing wall.
Also made naming conventions a bit more universal. ResetAll() -> Reset()

Random Bug fixes from 6months ago
Check if stuff is null before you try to use it. I think it's from taking
damage from weird places like exploding jellies that no longer exist or
pathes of fire on the ground that doesn't have a body? Just guesses
really.

Things not in this update
Passable UI