Winner Name Label
When a task is completed, put the name of who completed it over the task.
Needed to change what data is passed to the clients. Pass the name as well
as the task number
It only worked for displaying your name when you won. At the last second I
made it work when anyone completes a task instead. It'll probably work.
It's untested.

CreateDummyProBar
Testing to see if I can fit multiple progress bars easily. I can place
them, but it isn't easy to space them out properly.

Bug
Found abug when you exit the game. It probably only happens when you quit
to desktop. So then it's probably not an issue.

Really felt like I did more
I guess changing 1 UI thing at a time until it works takes a long time,
and then deleting the stuff that was unneeded makes it look like I didn't
do anything.