FarKill
FarKill works
Adjusted the distance needed. 100 is a bit too far. 70 is proabably also
too far, but oh well.
Added a progress meter, but forgot to test it.

Remember the last commit where I said I changed something last minute and
it would probably work?
Well, it did and it didn't.
I did create a game-breaking bug, BUT it wasn't in the spot I was thinking
it might be.
Apparently string.remove(15) doesn't like it if you give it a string
that's shorter than 15.
And remove(length) gives error and remove(length-1) removes the last
character of the name (was hoping if remove equalled the length, it would
just not remove anything.
Changed remove() to substring() and it doesn't break anymore
I wonder if I made the min and max close to 18, and set NoWrap, I probably
wouldn't have to cut off at 15. It might still be too long. And then I
would have to adjust the width to make it shorter which I don't know how
to do. At least I didn't see the option for it the first time I looked.