Update to work with Patch. UI Broken
Needed to add some dll files as dependencies to get it to compile
I think the game works now. Just the UI is the problem.

UI
Is still broken
HUD.objectivePanelController no longer exists so needed to find the panel
some other way.
Popups changed a bit. Needed to find the object that holds the text in a
different way. It used to be a direct child, now it's a grandchild.
Popups don't go away now.
Icons are still broken. They don't show up on the task list so you don't
know what item you'll get. I changed the loading to the legacy loading,
but it didn't help.
The offsets also seem a bit broken. Hard to really tell with no icons,
however.
Unsure if it works when you clear a level and go to the next. Need to test
it.