Fixed Some UI
Notifications are in the right spot, right scale and dissappear

Name of who completed the task is parented correctly, centered, and the
right size now

Item icons still don't work.
The command icon does work, however.
What is the difference?
command is a sprite
drones and items are textures
When using the inspector mod, the fields are populated.