Item Icons
Made item icons show up.
Had to do it my own way as opposed to just copying the version the
inventory uses.
Create an object, give it an image and text. Images use sprites instead of
RawImages that use textures. ItemDef has sprites defined so it works fine.

Also works for drones. Have to create the sprite from the texture.

Seems to work on stage 2. Stuff gets deleted and remade properly.

Still a little buggy.
Might not have everything quite in the right places.
And there might be an issue with a second icon when you complete a task.
Not sure why.