Failure Conditions
Mayron had some bugs at this point.
Client couldn't see the task list UI.
I tested and the client could see the UI.
I can't figure out what caused the bugs. I think it might be conflicts
with another mod.
Added some extra logging to get some more information.

BabyDrone
In a solo game, if your drone dies, CompleTask(-1) is the failure state.
It finished the task with the name being FAILURE so no one gets an item.

FarthestAway
Some extra debugging
Bug? When testing multiplayer, I think the host travelled farther away,
but I(client) won.

FewElites
Added some failure state code.

StartTeleporter
Failure code
Solo: if you complete this task, get rid of the progress bar. I had it be
full when you completed the task, but it was too noisy on the UI.

TasksPlugin
Version 1.2.0
RectWasNull - using these flags to reduce the amount of spam these checks
output to the log file
Added some more logging to try to catch some errors Mayron had. It's
mostly extra logging with no changes to the code.
GetPlayerCharacterMaster()
if you give it an invalid array index, it just returns the first element.
This is maybe naive. It might be better for this to just fail and null
reference exception than try to fix it in this way.
GetPlayerName(-1)
Pass in -1 to get the FAILURE name. This isn't maybe the best way to do
this, but it is an easy way.
Some safeties added so the -1 error case doesn't give items or try to send
a in message to the non-existant FAILURE player.

Readme
Added a wiki reference.
Made a wiki on github for the mod (not part of a commit).
It explains the tasks in more depth and how rewards work. Probably no one
will read it, but it's there and now I kinda know how to make a wiki.

Adjusted some spacing to hopefully look nicer.
Added version info. This version updated 3 tasks and added 1
newish(FewElites) one.